teachers/serializers.py
from .models import Teacher
from rest_framework import serializers, status

class TeachersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Teacher
        fields = '__all__'
        depth = 1