requirements.txt
pyjwt
django
djangorestframework
rest-framework-generic-relations
django-cors-headers
djangorestframework-simplejwt
Pygments
markdown
django-filter
gunicorn
django-extensions
django-avatar
django-allauth
social-auth-app-django
django-crispy-forms
django-dynamic-formsets
django-bootstrap4
psycopg2-binary==2.8.5
django-admin-sortable2
django-tinymce
django-ckeditor
django-rest-framework-social-oauth2
django-rest-knox
drf-nested-routers
drf-yasg
django-model-utils==4.0.0
celery
django-debug-toolbar
django-extensions
tldextract
boto3
django_cron
pyTelegramBotAPI
django-summernote
django-sortedm2m
django-admin-csvexport
django-modelclone